requirements.txt
streamlit
pandas
numpy
gspread
oauth2client
matplotlib
plotly
scipy
statsmodels
xlsxwriter
selenium
webdriver-manager
fake-useragent
openpyxl
scikit-learn
cachetools
altair
requests
watchdog
urllib3
SQLAlchemy
pymysql
bcrypt
python-dotenv
cryptography

extra-streamlit-components
pyjwt
